Ensure logfire.testing doesn't depend on pydantic and eval_type_backport
